requirements.txt
gradio==5.34.0
pillow
numpy
fastapi
uvicorn
diffusers
torch
transformers
